+++ pysys/utils/logutils.py
@@ -67,7 +67,7 @@ class BaseLogFormatter(logging.Formatter):
 		super(BaseLogFormatter, self).__init__(
 			propertiesDict.pop('messagefmt', DEFAULT_FORMAT),
 			propertiesDict.pop('datefmt', None) )
-		
+		assert not isinstance(self._fmt, binary_type), 'message format must be a unicode not a byte string otherwise % arg formatting will not work consistently'
 		if propertiesDict: raise Exception('Unknown formatter option(s) specified: %s'%', '.join(list(propertiesDict.keys())))
 
 
@@ -136,7 +136,20 @@ class ColorLogFormatter(BaseLogFormatter):
 		if os.getenv('PYSYS_COLOR',None):
 			self.color = os.getenv('PYSYS_COLOR').lower() == 'true'
 		
-		if self.color: self.initColoringLibrary()
+		if self.color: 
+			# initColoringLibrary, which might result in sys.stdout getting rewritten; PySys needs control of sys.stdout 
+			# so capture the new stdout then restore the original
+			stdoutbak = sys.stdout
+			try:
+				sys.stdout = sys.__stdout__
+				self.initColoringLibrary()
+			finally:
+				updatedstdout = sys.stdout
+				sys.stdout = stdoutbak
+			# now update the stream used by our handler
+			assert stdoutHandler.stream # unicodewrapper
+			assert stdoutHandler.stream.stream # actual stream
+			stdoutHandler.stream.stream = updatedstdout
 
 		super(ColorLogFormatter, self).__init__(propertiesDict)
 
@@ -227,10 +240,7 @@ class ColorLogFormatter(BaseLogFormatter):
 			try:
 				import colorama
 				colorama.init()
+				logging.getLogger('pysys.utils.logutils').debug('Successfully initialized the coloring library')
 			except Exception as e:
 				logging.getLogger('pysys.utils.logutils').debug('Failed to load coloring library: %s', repr(e))
 			
-		# since sys.stdout may be been redirected using the above, we need to change the 
-		# stream that our handler points at
-		assert stdoutHandler.stream
-		stdoutHandler.stream = sys.stdout